move websiteType to search package
