+++ backend/search_module/search/website_type.go
@@ -1,4 +1,4 @@
-package website_type
+package search
 
 import (
 	"encoding/json"
